test: parse_run_details: fix failing tests
This test has been failing since 66fe297 (WIP: first shot at condition
number, 2020-07-20).

Based on the default values of the OutputTables and OutputFilesUsed
set in NewRunDetails(), I think []string{} is the correct expectation
for both theses fields.  For the CpuTime field, I'm not sure, because
it's specified in the struct but not NewRunDetails(), which I would
have thought would lead to an error.  (The next commit will add
CpuTime to NewRunDetails().)