fix: ParseRunDetails: parse NumberOfDataRecords for $INFN output
When $INFN is used to subset the data set, the lst files lacks the
"NO. OF DATA RECS IN DATA SET:", but it does have a "TOT. NO. OF DATA
RECS:" line, so look for that instead.

This assumes that when $INFN is used there will be only one
"TOT. NO. OF DATA RECS:" present in the output and that it is a
reasonable value for NumberOfDataRecords.

Closes #227.