params: group parameter names when displaying multiple models
When --dir is used, the parameters are combined across models, and the
order of the de-duplicated parameter names is left as is.  This can
lead to columns with the same base name (e.g., "THETA8" and "THETA9"
from the golden file modified in this commit) being separated by
columns for parameters with different base names.

Callers will likely expect the columns to be grouped by base name, so
add logic for sorting the parameter names that puts all parameters
with the same base name name next to each other.

Re: https://github.com/metrumresearchgroup/bbi/pull/220#discussion_r661539150