feat: nmparser: record ONLYSIM in the run details
Several spots in bbi assume that NONMEM output includes estimation
details, but this isn't the case with ONLYSIM.  In preparation for
guarding these spots, add a field to RunDetails that indicates whether
ONLYSIM was used.

Note that using omitempty flag is important so that this addition is
compatible with the current bbr test suite.