fix: summary: consider ONLYSIM
Several spots in the `bbi nonmem summary` code path assume that there
are estimation methods in the output.  This isn't the case when
ONLYSIM is used, leading to out-of-bounds errors in a few places as
well as an inappropriate error about an empty .ext file.

Make use of the OnlySim field added a couple of commits back to guard
the estimation-specific code.

Note that it's not sufficient to just look at whether the list of
parsed estimation methods is empty because that can't be distinguished
from a partial .lst file, and bbr already has logic that checks the
latter.  This shows up as a test failure (BBR-SUM-004) if

  len(results.RunDetails.EstimationMethods) == 0

is used instead of OnlySim.

Fixes #224.