fix: nmparser: accept ONLY alias for ONLYSIMULATION
When ParseRunDetails() encounters a line beginning with $SIM, it sets
.OnlySim according to whether it sees "ONLYSIM" in the line.  That
covers the option's "ONLYSIM" and "ONLYSIMULATION" spellings, but the
option can also be shortened to "ONLY".

No $SIM options aside from ONLYSIMULATION include "ONLY" as a
substring, so adjust ParseRunDetails() to look for "ONLY" rather than
"ONLYSIM".  Also extend the ParseRunDetails test to cover more
variants.