nmparser: use strToFloat to handle NaN objective function values
The following bbr error was reported for a model:

  > model_summary(mod)
  Error: parse error: premature EOF

                       (right here) ------^

The root cause was NaN values for the objective function in the lst
file:

  OBJECTIVE FUNCTION VALUE WITHOUT CONSTANT:                        NaN
  OBJECTIVE FUNCTION VALUE WITH CONSTANT:                           NaN

An earlier commit in this series makes the failure more obvious
because bbr will report that the bbi call failed and include

  level=fatal msg="json: unsupported value: NaN"

in the output.

Parse the objective function values with strToFloat() so that NaN's
are mapped to DefaultFloat64, allowing them to be encoded to JSON and
handled downstream by bbr.

Fixes #266.