nmparser: map NaN's to DefaultFloat64 too
The primary way to send results out is via JSON, and NaN values can't
be encoded, so map them to the same default value that strToFloat()
uses for errors.

This loses the distinction between an error and a NaN value, but 1)
it's better than leaving an NaN value that crashes with an encoding
error and 2) bbr already detects DefaultFloat64 and maps it to
NA_real_, so the distinction is already being lost in the other
direction.