nmparser: support multiple values for estimation and covariance time
If there are multiple estimation methods, there are multiple
estimation and covariance times.  Report them all as a list so that
consumers (in particular bbr) can access these values.

This assumes that, when there are multiple estimation methods, 1)
these times will be encountered in the same order as the
estimation_method values, 2) no method has multiple times for
estimation or covariance, and 3) no method is missing a time for
either estimation or covariance.

I think the first and second are safe assumptions.  The third one
seems to hold for the models in integration/, though I'm not sure that
it holds in general.  However, for cases where the third assumption is
violated, the failure mode seems low risk and acceptable: the consumer
will be unable to map times to a particular method.

Re: https://github.com/metrumresearchgroup/bbr/issues/509