summary: extract repeated GetModelOutput call to function
The summary command has two branches, one for when a single model is
specified and one for multiple models.  Both use the same
GetModelOutput() call, with the latter calling it on each model.

The code isn't too complicated, but it's still worth pulling it out
into a helper so that it is clearer at a glance that these two
branches are using the same call.