summary: improve error signaling for --json
The summary command doesn't reliably relay failures when --json is
specified and multiple models are given.  The main issues demonstrated
by TestSummaryPathMultipleModelsError test added a couple of commits
back are that 1) by default errors are not present in the JSON output
and 2) in cases where they are, they are split off into a separate
array, preventing the consumer from identifying the model that failed
by position.

Fix these two issues by embedding a result's status and error message
in the parser.SummaryOutput struct.  For a given result, a consumer
can know if it succeeded by looking at the boolean success field.

Also, keep the top-level "Errors" array but, instead of the Error
values, record where any failures are positioned in the "Results"
array.  This allows the consumer to determine whether there are any
failures without checking every result and to easily extract the
failed records.

Note that there is very similar handling/reporting of multiple
modeling in in params.go.  At some point, its JSON error should be
updated too.