summary: improve consistency of non-zero exits
If summarizing any model fails, the summary command exits with a
non-zero status when 1) --json is not specified or 2) --json is
specified and a single model is given.  As of the previous commit,
error messages should reliably make it into the JSON output, so the
caller doesn't _need_ the exit code to know that there was a failure.
Still, it'd be good to use a non-zero exit in all cases for
consistency.

Drop the early return from the --json code path for multiple models so
that it goes through the same "error -> non-zero exit" handling that's
done for the "no --json, multiple models" code path.