integration/execute.go: disable "receives same value"" linter check
When a few commits back added an executeCommandNoErrorCheck() call to
integration/bbi_summary_test.go, that led to golangci-lint giving the
following warning:

  integration/execute.go:42:54: `executeCommandNoErrorCheck` -
  `command` always receives `"bbi"` (unparam)

It wants us to drop the command parameter and hard code "bbi" in the
function.  The warning presumably shows up now because the number of
executeCommandNoErrorCheck calls reached some threshold.

The more general form seems worth keeping around, even if we always
use "bbi" at this point, so squelch the warning.