nmparser: don't omit zero for times in run_details JSON output
RunDetails uses the omitempty tag for PostprocessTime and CpuTime.
However, if a value of zero is in the lst for either of these, that
should still be reported, so drop the tag.

Note that NewRunDetails() sets these values to DefaultFloat64, so
whether or not a value was parsed from the lst file can be
distinguished.

Closes #273.