reclean: avoid index error if no positional argument is specified
`bbi nonmem reclean` requires one positional argument and fails with
an out-of-bounds error if it doesn't get it:

  panic: runtime error: index out of range [0] with length 0

  goroutine 1 [running]:
  github.com/metrumresearchgroup/bbi/cmd.reclean(0xc0001a2000?,
  {0x1694b08, 0x0, 0x0?})
  [...]

Give a more appropriate failure by making reclean() return an error if
it doesn't receive a single positional argument.  Another option would
be to use cobra.ExactArg, but that leads to a less informative
message ("Error: accepts 1 arg(s), received 0").