nmparsers: add some checks for unexpected estimation method lengths
With an incomplete lst, `bbi nonmem summary` (without or without
--json) can fail with an error like this:

  $ head -n 50 integration/testdata/bbi_summary/acop/acop.lst | \
      sponge integration/testdata/bbi_summary/acop/acop.lst
  $ bbi nonmem summary integration/testdata/bbi_summary/acop/acop.lst
    FATA[0000] GetModelOutput: runtime error: index out of range [-1]
    [...]

That comes from CheckIfNotGradientBased(), which assumes that there is
at least one estimation method (which may not be the case for an
incomplete file).  Update CheckIfNotGradientBased() and
CheckIfBayesian() (which has the same problem) to guard against an
empty method array.

Once that is resolved, calling summary without --json hits into
another index error:

  $ bbi nonmem summary integration/testdata/bbi_summary/acop/acop.lst
  panic: runtime error: index out of range [-1]
  [...]

Fix that by teaching GetModelOutput() to return an error if, outside
of the ONLYSIM context, the parameter data length and estimation
method length don't match, taking that as a sign that the lst parsing
is incomplete.  Catching this issue at the level of GetModelOutput()
means that --json also shows this error rather than a misleading and
incomplete success=true result.

Thanks to @seth127 for providing the reproducer.