summary: catch panics from GetModelOutput
bbr's model_summaries() currently calls `bbi nonmem summary --json`
with one model at a time, catching any errors with tryCatch().  It's
going to move to passing multiple models, so it is important that
outputting JSON doesn't fail because of one bad model.  Ideally
GetModelOutput() wouldn't hit into panics and we could rely on it
returning an error that we can relay, but we know that's not the
current state.

Recover the panic and output its full backtrace.  Hopefully that will
make it clear that this is a bbi bug, and that it should be reported.