sge: use process bbi executable in scripts unless overridden
`bbi nonmem run sge` generates scripts that call the bbi executable
stored in the BbiBinary configuration field.  This defaults to the
executable of the current process and is overridden by the bbi_binary
value in bbi.yaml (which in turn can be overridden by the --bbi_binary
flag).  Notably calling `bbi init` writes the resolved executable to
bbi.yaml, so by default uses the bbi executable that was resolved at
the time `bbi init` was called.

As described in gh-261, this has some downsides:

 * bbi.yaml is typically tracked in version control, meaning that
   different users must either install bbi in the same location or
   override the path by making local modifications to bbi.yaml or
   passing bbi.yaml.

 * from the user perspective, it's not obvious that bbi_binary only
   affects SGE runs.  This discrepancy is particularly confusing for
   bbr users setting bbr.bbi_exec_path.

Address the above issues by 1) leaving BbiBinary unset by default,
leading to `bbi init` _not_ writing the value to bbi.yaml, and 2)
resolving the executable path at the time of SGE script generation
unless the bbi binary was explicitly specified (either via bbi.yaml or
--bbi_binary).

Note that the above approach only affects directories with a newly
generated bbi.yaml.  The treatment of already configured directories
remains the same.

Closes #261.