nonmem: also treat maxlim=0 as special case
As described in the previous commit, processNMFEOptions() interprets a
maxlim value of 100 to mean "don't pass a -maxlim argument to nmfe".
100 is the default value, so, aside from it ending up in bbi.yaml,
it's mostly an internal detail.

An upcoming commit is going to change the default to 2.  We'll then
need to advertise to the caller how to say "unset -maxlim".  0 seems
like a more intuitive value than 100, so give it the same treatment.