nonmem: tighten maxlim value special casing
nmfe's accepts 1, 2, and 3 as valid values for its -maxlim argument.
bbi's config.NMFEOptions.MaxLim defaults to 100, a value that
instructs processNMFEOptions() to not pass -maxlim to nmfe.
processNMFEOptions() takes anything below 50 as a valid -maxlim value.
This seems unnecessarily broad, and makes it possible to pass invalid
values to nmfe.

Tighten the condition so that

 1) only valid maxlim values are passed to nmfe
 2) 100 is treated as the special value that means "don't pass
    -maxlim"
 3) all other values lead to a warning