configlib: drop inaccurate MaxLim comment
ac48a49 (refactor: use snake_case per R/python userbase, 2020-02-16)
changed the NMFEOptions.MaxLim comment to say "Default (empty value)
is 3" rather than "Default (empty value) is 100".  That was incorrect
at the time, because the default was still 100, and it's incorrect now
that the default is 2.

There doesn't seem to be much value in mentioning the default here, so
just drop the comment entirely.