summary: sort Errors field
When multiple models are specified, the summary JSON output includes
an Errors field that lists positions of Results items that had an
error.  This list of ints unsorted.

That's okay from bbr's perspective: it depends only on the Results
elements being in the same order as the model arguments.  However,
it's still nice to have these come in a predictable order, especially
because we have tests that compare against golden files containing the
JSON output.  (I've yet to observe these failing intermittently,
though I think that's possible if these are left unsorted.)