GetModelOutput: return error if lst does not have $DATA line
In metrumresearchgroup/bbr#527 [*], Seth noted that `bbi nonmem
summary --json --no-ext-file` reports success=true for an lst file
with nonsense content.  This success=true behavior was introduced by
7c0caa76 (nmparsers: add some checks for unexpected estimation method
lengths, 2022-07-29).  Before this commit, the above command (with or
without --json and/or --no-ext-file flags) would have failed with an
index error in CheckIfNotGradientBased().  That commit guards against
that error and adds a check to abort when the ext data length and
method length don't match.

However, when --no-ext-file is specified, the length check isn't
triggered (there's no ext data to compare).  Without the --json flag,
SummaryOutput.Summary() will hit into an index error in this case.
With the --json flag, we get the incorrect success=true result.

Prevent the index error (no --json) and success=true result (--json)
by checking for a field that we expect to always exist in the lst
file, $DATA.  There are probably a good number of other things that
would make sense to check, and those can be added as needed.  However,
it's diminishing returns after the first check; checking _any_
should-always-be-present field catches the entirely invalid case,
while any additions would catch narrow windows in an incomplete run.

[*] https://github.com/metrumresearchgroup/bbr/pull/527#pullrequestreview-1058066983