integration: specify NONMEM version for NMQUAL tests separately
NMQUAL has been deprecated for NONMEM 7.5, and Metworx does not have
a *gf installation for 7.5 [*].  In preparation for updating Drone to
use NONMEM 7.5, convert the NMQUAL environment variable into a
variable that specifies both whether to run the NMQUAL tests and which
version of NONMEM to use.

Thanks to @seth127 for suggesting that the NMQUAL tests could use 74
while other tests use 75 rather than adding both 74 and 75 runs to the
Drone configuration.

[*] https://kb.metworx.com/Users/Tutorials/Nonmem-and-PsN/