integration: add and wire up AssertNonMemCleanedUpFiles
For the integration tests that call AssertNonMemCreatedOutputFiles(),
also call AssertNonMemCleanedUpFiles() to check that some files that
should have been cleaned up do not exist.

The only more involved change here is for TestNMFEOptionsEndInScript.
The wrinkle there is that the test 1) calls `bbi nonmem run local` and
2) then runs the generated script.  AssertNonMemCleanedUpFiles() needs
to be called _before_ the second step because bbi's cleanup won't be
in effect for a direct run of the script.  So, add a small set of
assertions in between step 1 and 2, including a call to
AssertNonMemCleanedUpFiles().