TestParseRunDetails: make input differences clearer
TestParseRunDetails() defines two long []string{} inputs where only
one line differs, and an upcoming commit will add another case.  This
approach makes it hard to spot what the distinction is.

Instead define a baseInput variable, copy it, and tweak just the line
of interest.  Using an index to pick out the line to modify isn't
ideal, but it's an improvement over repeating the entire slice.