summary: account for chain method in incomplete run check
As of the last release (specifically 7c0caa76), GetModelOutput()
compares the lengths of RunDetails.EstimationMethods and of the
ParseParamsExt() return value, and errors if they don't match.  The
intent is to catch incomplete runs, but this check is incorrectly
triggered for METHOD=CHAIN because that method does not have
corresponding estimation data.

Update the check so that the chain method is not counted toward the
expected length.  Based on the NONMEM help, it appears that all other
method types should have corresponding estimation values.

Closes #288.