drone: move built bbi's directory to front of PATH
.drone.yml builds bbi and copies it to /data/NNN/apps/, and puts that
directory at the end of $PATH.  That approach has worked okay for a
long time (first occurrence introduced in 3508b4b4, 2020-04-16)
because Metworx did not ship with a bbi in any of its PATH
directories.  That's no longer the case with the latest Metworx.

Move /data/NNN/apps/ to the front so that the built bbi takes
precedence.