init: make compatible with Windows
On Windows, 'bbi init --dir X' doesn't write "nonmem" entries to
bbi.yaml even if X is valid.  There are two issues:

 * isPathNonMemmy() and findNonMemBinary() assume that a nmfe binary
   will have an executable bit, but that isn't the case on Windows

 * findNonMemBinary() doesn't account for the nmfe binary ending with
   ".bat" on Windows

Fix isPathNonMemmy() by skipping the executable bit check on Windows.

For findNonMemBinary(), add a new function.  Another option would be
to update findNonMemBinary() with conditional logic but 1) it's nice
to have a separate function we can test when not on Windows
(especially given we don't have Windows CI right now) and 2) the
duplication isn't too bad because the remaining shared logic is mostly
just listing a directory.

Fixes #295.