run local: invoke bash directly rather than relying on shebangs
When `bbi nonmem run local ...` is called, a NONMEM command is written
to `{model}.sh`, and the script is passed as the first argument to
exec.Command(), relying on the presence of a /bin/bash shebang.
Likewise, if --post_work_executable is specified, the
post_processing.sh path is passed as the first argument.

On Metworx (and most Linux systems), hard coding /bin/bash works fine,
but it will fail on systems that happen to have bash at a different
location.  To avoid hard coding the path, we could find the bash
executable and write the appropriate shebang.  However, that won't
work for Windows, which doesn't support shebangs by default.  So
instead pass bash as the main program to exec.Command(), feeding the
script as the argument.

This still won't work _by default_ on Windows systems.  But

 * there's now a clearer failure saying that bash is required

 * it is relatively straightforward for users to install bash via
   either Rtools or Git for Windows (and many users will already have
   one of these available).  And then in the context of R/bbr, users
   will need to adjust their .Renviron with an entry like

     PATH="${RTOOLS40_HOME}\usr\bin;${PATH}"
     # Or
     # PATH="C:\Program Files\Git\bin;${PATH}"

Note the shebangs are still included in the scripts and the scripts
are still executable, so if people were used to being able to debug by
going to the model directory and running `./x.sh`, they can still do
so.  Also, note that executing the SGE grid.sh script still relies on
the hard-coded /bin/bash shebang.  (We could adjust the qsub
invocation to avoid that, but it's not likely to matter in practice.)