run: quote tokens if needed when generating shell scripts
`bbi nonmem run ...` generates a {run}.sh bash script and, for sge,
also generates grid.sh.  Both of these scripts have parts that may
need quoting:

 * grid.sh: the value of the -wd directive, the path to the bbi
   executable, and the name of the run

 * {run}.sh: the path to the nmfe executable and the name of the run

Although it's not hard to hit into a case that needs quoting on *nix
systems, Windows users are especially likely to encounter this problem
due to '\' in paths (e.g., in the nonmem home value that 'bbi init'
writes to in bbi.yaml).

To handle these cases, define a ShQuote() helper that checks if a
token contains a character that needs quoted and if so surrounds it
with single quotes.  This is the same quoting approach used by
shlex.quote() in Python's standard library.

Note, however, even if we quote the values, some cases still trip up
downstream code outside of bbi:

 * NONMEM's nmfe scripts do not handle spaces in the model name.

   $ /opt/NONMEM/nm75/run/nmfe75 'a b.ctl'  'a b'.lst  -maxlim=2
   + /opt/NONMEM/nm75/util/OPTION_CHECK a b.ctl a b.lst -maxlim=2
    Option not recognized: a
    Option not recognized: b.lst
   + status=2

 * nmqual/autolog.pl triggers a nmfe call that isn't quoted when the
   model name needs quoted, and, if the leading path needs quoted, a
   chdir in nmqual/autolog.pl fails (at least in the case of a path
   with spaces).  So, the quoting added here doesn't make any cases
   work, but it at least leads to easier to digest failures in some
   cases (e.g., the leading path with spaces fails with a message
   about being unable to chdir rather than silently aborting with an
   exit code of 0).

Also, there are cases that work locally but fail for SGE submission
both before and after this commit.  For example, the new "modquote"
test case works (after this commit) for the local submission, but with
'run sge' ends up in an Eqw state.  Entering the directory and running
the grid.sh file completes successfully.  It seems likely that SGQ
isn't satisfied with the -wd directive quoting, but 1) as the
"leading-path-with space" case shows, the -wd value does need to be
quoted, and 2) I haven't found a broken case that isn't also broken
before this change.

In the long run, it's worth considering reworking bbi's approach to so
that we don't rely on generating bash scripts with arbitrary paths.
For example, for the {run}.sh case, we could instead write a file in
the output directory with the nmfe executable path and the arguments
(no quotes needed) and then have executeLocalJob() read those and use
os/exec to call nmfe, avoiding the shell.