init: generate NonMemDetail map directly
initializer(), the function hooked up as the init command's RunE,
loops over the NONMEM installation locations and uses viper.Set() to
create each value of the "nonmem" configuration section.

For more direct testing, pull this logic outside the RunE entry point,
allowing the tests to pass in the directories and inspect an
intermediate structure.  One option would be to have this new function
return the Config object, but instead keep the return value focused on
the only thing that may be changed, the .Nonmem entries.

Note that this is intended to be functionally equivalent to the
previous approach, including _not_ being affected by an existing
bbi.yaml (i.e. any non-default configuration values are discarded).