init: warn on identifier collisions
'bbi init' allows specifying multiple directories by passing --dir
multiple times.  The identifier for the section key under "nonmem" is
taken from the basename, so it's possible for there to be a collision
in the identifier.

The current behavior is to select one entry.  It's last one wins and
based on the order the directory is in, so which particular entry wins
is undefined.  Even though hitting into this is unlikely, changing
this to an error is a bit too aggressive; it probably won't be clear
to the caller how to address it, and they won't have the ability to
modify the NONMEM install directory in many cases.

Instead give a warning.  If the user sees the warning and decides that
want they discard entry, they have the written entries to use as a
guide.