init: sanitize config identifier for nonmem entries
Viper keys cannot contain a period because that's the delimiter (by
default).  Until a few commits back when 'bbi init' stopped using
viper.Set(), a directory with a period wouldn't even produce a valid
entry.  For example, "7.5.0" would produce

   nonmem:
        7: {}

Following that commit, 7.5.0 will make it through as a key.  However,
downstream access of that value in nonmem.go will fail.

Sanitize directory names by replacing periods with hyphens.  In the
unlikely case that doing so leads to a collision, emit a warning.