local: add ignore pattern for WK_* files
filesToCleanup() does not mark a WK_ buffer file for cleanup if it
isn't empty because that indicates that the user probably wants to
make adjustments so that everything fits into memory.  However, a
buffer file isn't something you'd want to track in Git.

Although it's possible that this entry collides with a user output
file (e.g., a table name), it seems reasonable to treat WK_ as a
reserved prefix that users should avoid given that NONMEM uses it for
its worker files.