nonmem: add empty WK_ files to list of files to clean up
Under --parallel, NONMEM creates several WK_* files.  These are worker
files created for parallelization problems and are used as file
buffers.  The NONMEM Users Guide Introduction to NONMEM 7.5.0 (pg 73)
says

  One can alternatively assess empirically whether file buffers are
  used, by beginning the run, allowing perhaps one iteration to
  transpire, then from another command window do a directory search
  for FILE*, (or WK* for worker files in parallelization problems,
  section I.72 Parallel Computing (NM72)

  If any of the FILExx do not have 0 size, then they are being
  used. Interrupt the analysis, then increase the appropriate LIM
  value with the $SIZES record [...]

So, a non-empty file is a signal that the user probably wants to make
adjustments so that everything fits into memory.  An empty file, on
the other hand, is safe for us to clean up [*].

Due to the regexp match and the extra size condition, this doesn't
work nicely as part of getCleanableFileList().  filesToCleanup()
already has special handling for some parallelization, so add the WK
handling there.

[*] NONMEM cleans up FILE* buffer files after a run.  I'm not sure why
    it doesn't also clean up WK*.

Re: #285