bbi_local_test.go: set up config for each scenario
TestSpecifiedConfigByAbsPathLoaded sets up a bbi.yaml file under
$ROOT_EXECUTION_DIR/meow (within an anonymous function) and then loops
over several scenarios, passing this config file to bbi's --config.

Now that the setup has been extracted to copyConfig(), it's clearer
and more isolated to just do the (inexpensive) setup for each
scenario.  And with the next commit, copyConfig() will return the path
to the config; inlining the copyConfig() call will avoid needing to
relay the return value across an additional layer.