bbi_local_test.go: fix deferred close of files
When copying the config for the --config tests, the source and
destination files are closed via

    defer t.R.NoError(f.Close())

That doesn't behave as intended because argument evaluation is not
deferred.  The downstream io.Copy() then uses closed files, leading to
a "copy_file_range: use of closed file" error.

That error is not checked, so the tests continue with an empty config
file.  Loading an empty config still shows the message that
AssertSpecifiedConfigLoaded() is looking for, so the tests pass
despite the actual NONMEM call failing with

    nmVersion of  was provided but has no configurations in bbi.yaml!

To make this test behave as intended

 * don't check the source file's .Close() error because it's not
   essential that a read-only file handle closed without error

 * delay the destination .Close() check by wrapping it in a function