go.mod: bump minimum go version to 1.16
The tests started using os.WriteFile in e09a001 (init: make compatible
with Windows, 2023-02-02), but it isn't available until 1.16.  We're
due for an update of the minimum Go version, but for now just increase
to 1.16.

As of Go 1.16, ioutil.{ReadFile,WriteFile,ReadFile} are deprecated and
just relay functions in io or os.  Use the new functions directly.

This issue was flagged by the upcoming lint GitHub Actions that runs
with the minimum Go version specified in go.mod.