ci: switch to GitHub Actions
For bbi testing with Drone, we relied on a self-hosted server with
NONMEM in order to run the full test suite.  At least for now, the
plan isn't to mimic that setup with GitHub Actions.  Instead set up
GitHub Actions to run all the tests that don't require NONMEM, and
rely on the full test suite being executed outside of CI.