nmparser: delete parse_parameter_comments.go
This file was added 258b6906 (work on parsing out parameter names,
2016-03-16), and its ParseParameterNames was called by
ParseLstEstimationFile.  That call was replaced with to-do comment in
2bb8d9c7 (wip: start cleanup to support bayesian and models where
parameter structures/names aren't parseable, 2019-11-27).

Remove the file.  If the code is later useful when implementing the
to-do, it's easy enough to find which commit added that comment and
resurrect parse_parameter_comments.go.