cmd/probs: fix handling of optional directory argument
If a positional argument is passed to 'bbi probs', the command looks
there for model files.  Otherwise it falls back to using the current
working directory.

The handling of the argument is incorrect unless it happens to point
to the current working directory.  The model files come back relative
from ListModels, so modSummaries will look in the current directory
for them despite the directory being specified.

Fix the error by making the files absolute before passing them to
modSummaries.