integration: rename postrun to nmless
We recently split the integration package into the tests that require
NONMEM (now under "nonmem") and those that don't (under "postrun").
"postrun" wasn't a great name choice because the key aspect of these
tests are that they don't require NONMEM, not that they happen after a
NONMEM run.

Rename it because some upcoming tests aren't going to fit the
"postrun" description.