clean: fix handling of files from *_copied.json
If the --copiedRuns argument is used, clean grabs the files to remove
from the *_copied.json file set up when a run uses --copy_lvl.
*_copied.json records the files as absolute paths.

clean doesn't process absolute paths correctly because it
unconditionally prepends a directory, leading to an invalid path.  As
a result, the files don't get removed.

Update clean to only append dir if the file isn't an absolute path.