goreleaser: update config format to version 2
The goreleaser action set up in 1449d16a (ci: switch to GitHub
Actions, 2024-08-09) requires version 2 of the format.

Note that there is already a "builds" section that is compatible with
version 2.  Presumably the "build" section being removed by this
commit had no effect in the version of goreleaser used in the older
Drone build.

According to 'goreleaser check', this is config is now valid.