ci: set token for goreleaser to push to homebrew-tap
The builtin GITHUB_TOKEN is sufficient for goreleaser to make a
release in _this_ repository, but a custom token is needed in order to
push to a different repository than the one workflow is running in
(metrumresearchgroup/homebrew-tap in this case).

Set GH_GORELEASER_BREWS_TOKEN in the environment.  (The previous
commit updated the brew's section of the goreleaser config to use
this.)