[todo msg] sge: convert qsub options to template directives
This has the advantage capturing more of details about the run on
disk.

And now that submission happens with a plain 'qsub $script` call, it
opens up the possibility of allowing the callers to override the
template (e.g., via a command-line argument).

[todo] discuss quoting
e71dc47f (run: quote tokens if needed when generating shell scripts, 2023-02-15)