run-integration-tests: adjust MPIEXEC_PATH for Metworx compatibility
In the upcoming Metworx, /usr/bin/mpiexec won't exist.

We could set MPIEXEC_PATH to the default value of
/usr/local/mpich3/bin/mpiexec (or even drop the handling entirely).
The main downside is that tests could then no longer distinguish
whether overriding the default mpiexec path with --mpi_exec_path was
effective.

Set this to a value that 1) works across Metworx versions and 2) is
distinct from the default bbi value by resolving the default location,
which we can rely on existing and being a symlink.  (Another option
would be to set it to the highest precedence mpiexec in PATH.)

While at it, make it possible to choose a different location by
setting MPIEXEC_PATH in the calling environment.