sge, local: decouple script templates
nonMemExecutionTemplate is used by sge for writing the grid.sh script
with the 'bbi nonmem run local ...' command and by local for writing
the {run}.sh script with the nmfe command.  local's use works, though
it's a bit odd that it includes a "#$ -wd ..." comment directive that
is only relevant for the SGE submission.

The next commit will add even more directives and a later commit will
add another template for Slurm.  Replace nonMemExecutionTemplate with
a dedicated template for local and SGE so that the local template
doesn't contain several noop SGE directives.