executeSGEJob: log output on ExitError
Like executeLocalJob, executeSGEJob runs a command with CombinedOutput
and returns a turnstile.ConcurrentError if that results in an error.
Unlike executeLocalJob, it doesn't relay the combined stdout/stderr
when it hits an ExitError.  The ConcurrentError doesn't include the
stdout/stderr, and, in the error case, the output isn't written to
disk, so the caller has no way to access these details.

Update executeSGEJob to log ExitErrors in the same way executeLocalJob
does.