runModelCommand: write output file even on failure
The new runModelCommand was extracted from executeLocalJob and
executeSGEJob and behaves as they did in terms of handling the output.
The output is collected with CombinedOutput.  If there's an error, the
output is logged before returning a ConcurrentError.  The output is
written to _disk_ only if there was not an error.

Capturing the output on disk is helpful for troubleshooting.  Rework
runModelCommand to send standard output and standard error directly to
a file.