sge: convert qsub options to template directives
This has the advantage of capturing more of details about the run on
disk and will also make it easier to align the submission logic across
different batch systems because the call is reduced to a submission
command that takes the submission script as the positional argument.

And now that submission happens with a plain 'qsub {script}' call, it
opens up the possibility of allowing callers to override the template
(e.g., via a command-line argument).