local, sge: do shell quoting at template level
For local and SGE execution, various spots in the generated shell
scripts may need quoting.  The ShQuote helper [*] is selectively
applied to the working directory (for SGE the directive) and to
different pieces of a command before joining them into the final
command string.

The next commit will move all the qsub options into the SGE template.
That leads to more values that may need to be quoted (because they are
coming through at the script level rather than via exec.Command
arguments).

We could continue to handle all of the quoting outside the template,
but it's a better fit for the template because what needs quoted is
inherently tied to how the script is rendered (which is the
responsibility of the template to define).  Also, in the future we may
allow users to override the template, in which case they should be
able to control quoting rather than relying on some upstream and
inaccessible code to quote the right values.

Note that this new approach sends all of the command's pieces through
ShQuote.  That's relatively inexpensive and won't lead to superfluous
quoting because ShQuote first checks whether the item needs quoting.

[*] ShQuote was added in e71dc47f (run: quote tokens if needed when
    generating shell scripts, 2023-02-15).  See that commit's message
    for an overview of the quoting situation.